Implmented TopKFrequent selection using a min heap
This was quite intersting to get my head around,
and I also feel my imlmentation is quite messy :(

The idea is to maintain a min heap of length k,
allowing us to easily compare each item with the head
if the heap and swap it if it is larger - then bubble it into
the correct position.

Interestingly (and shouldnt be too much of a suprise)
this is not magnitudes better than my previous linq solution
when submitting on leetcode. linq's `.OrderByDescending()`
and underpinning OrderedEnumerable perform about the
same and is significantly more readable.